move home page into separate file
